Move integration into packages
following the gbm patten
